+++ scripts/module.ts
@@ -1,21 +1,23 @@
 /// <reference types="../node_modules/fvtt-types/src/index.d.mts" />
 /// <reference path="types/project-fu.d.ts" />
 
-import { Logger } from './utilities/logger.js'
 import { HookManager } from './services/hook-manager.js'
 import { EffectDialog } from './ui/effect-dialog.js'
+import { Interpreter } from './services/interpreter.js'
+import { initializeSocket, registerSocketHandler } from './utilities/socket-management.js'
+import { UpdateRequest } from './types/types.js'
 
-Hooks.once('socketlib.ready' as any, () => {
-  EffectDialog.initializeSocket()
-})
+function setupSocket() {
+  initializeSocket()
+  registerSocketHandler('showEffectDialog', (requests: UpdateRequest[]) => EffectDialog.showDialogForUser(requests) )
+  registerSocketHandler('processEffect', (request: UpdateRequest, sourceUuid: string, itemId: string, targetUuids: string[]) => 
+    Interpreter.processEffectAsGM(request, sourceUuid, itemId, targetUuids)
+  )
+}
+
+Hooks.once('socketlib.ready' as any, () => { setupSocket() })
 
 Hooks.once('ready', async function() {
   if (game.system?.id !== 'projectfu') return
-
-  if (typeof (globalThis as any).socketlib !== 'undefined') {
-    EffectDialog.initializeSocket()
-  }
-
-  const hookManager = HookManager.getInstance()
-  hookManager.initialize()
+  HookManager.initialize()
 })