+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [1.0.2] - 2025-09-20
+
+### Added
+- Spell and skill support
+- HTML parsing safety
+
+### Improved
+- Socket management 
+- Amount parsing by using the system expressions
+- Object types
+- Code structure
+- Linting
+
+### Fixed
+- Bug in resource pipelines
+- Parsing of resource event args containing parenthesis; for step expressions
+- Proper cleanup of hooks on module disable
+
 ## [1.0.1] - 2025-09-17
 
 + initial release